requirements.txt
cryptography==41.0.3
Jinja2==3.1.2
pycryptodomex==3.19.1
